+++ backend/src/main/java/com/sirheadless/mybookmarks/url/controller/UrlController.java
@@ -2,6 +2,8 @@ package com.sirheadless.mybookmarks.url.controller;
 
 import java.util.List;
 
+import com.sirheadless.mybookmarks.authenticationfacade.AuthenticationFacade;
+import com.sirheadless.mybookmarks.category.entity.Category;
 import com.sirheadless.mybookmarks.url.entity.Url;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -26,6 +28,10 @@ import com.sirheadless.mybookmarks.url.service.UrlService;
 @Controller
 @RequestMapping("url")
 public class UrlController {
+	
+    @Autowired
+    private AuthenticationFacade authenticationFacade;
+    
 	@Autowired
 	private UrlService urlService;
 	
@@ -36,20 +42,33 @@ public class UrlController {
 		return new ResponseEntity<Url>(url, HttpStatus.OK);
 	}
 	
-	@GetMapping("/")
+	@GetMapping("")
 	public ResponseEntity<List<Url>> getAllUrls() {
 		List<Url> list = urlService.getAllUrls();
 		return new ResponseEntity<List<Url>>(list, HttpStatus.OK);
 	}
+	
+	@GetMapping("all/{userId}")
+	public ResponseEntity<List<Url>> getAllUrlsByUserId(@PathVariable("userId") Integer userId) {
+		System.out.println("UrlController get all Urls for userId: " + userId);
+		List<Url> list = urlService.getAllUrlsByUserId(userId);
+		return new ResponseEntity<List<Url>>(list, HttpStatus.OK);
+	}
+	
 	@PostMapping("add")
 	public ResponseEntity<Void> addArticle(@RequestBody Url url, UriComponentsBuilder builder) {
-        boolean flag = urlService.addUrl(url);
-        if (flag == false) {
-        	return new ResponseEntity<Void>(HttpStatus.CONFLICT);
-        }
-        HttpHeaders headers = new HttpHeaders();
-        headers.setLocation(builder.path("/article/{id}").buildAndExpand(url.getUrlId()).toUri());
-        return new ResponseEntity<Void>(headers, HttpStatus.CREATED);
+		Integer userId = authenticationFacade.getUserId();
+		if(userId != null) {
+	        boolean flag = urlService.addUrl(url);
+	        if (flag == false) {
+	        	return new ResponseEntity<Void>(HttpStatus.CONFLICT);
+	        }
+	        HttpHeaders headers = new HttpHeaders();
+//        	headers.setLocation(builder.path("/article/{id}").buildAndExpand(url.getUrlId()).toUri());
+	        return new ResponseEntity<Void>(headers, HttpStatus.CREATED);
+		}
+		return new ResponseEntity<Void>(HttpStatus.INTERNAL_SERVER_ERROR);
+		
 	}
 //	@PutMapping("update")
 //	public ResponseEntity<Url> updateArticle(@RequestBody Url url) {
@@ -58,8 +77,28 @@ public class UrlController {
 //	}
 	@DeleteMapping("delete/{id}")
 	public ResponseEntity<Void> deleteArticle(@PathVariable("id") Integer id) {
+		
+		ResponseEntity<Void> noAuthentication = checkAuthentication(id);
+		if (noAuthentication != null) {
+			return noAuthentication;
+		}
+		
 		urlService.deleteUrl(id);
 		return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
 	}
 	
+	private ResponseEntity<Void> checkAuthentication(int id){
+		Integer userId = authenticationFacade.getUserId();
+		if(userId != null) {
+			return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
+		}
+		Url url = urlService.getUrlById(id);
+		if (url == null) {
+			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+		}
+		if (url.getUserId() != userId) {
+			return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
+		}
+		return null;
+	}
 }
